Android: fix custom dimensions reporting
Fix to custom dimension reporting so that it is not undertaken
separately to analytics tracking, but part of the original tracking
item/event. When using multiple dimensions, the code was sending a hit
per custom dimension, this incorrectly resulted in multiple hits being
recorded for the one screen view/event/etc.

Also updated the customDimensions HashMap to have the key as an integer
rather than to rely on parseInt later when adding the dimension to the
tracker.